Armature Export (preliminary)
Added preliminary armature export.  This has been tested on a basic
level to work with both Unreal Engine 4 and Unity 5.  Currently doesn’t
support animation filtering and the option to not export it, as well as
a few other things.  The Animation export preferences currently do
nothing, but will do something in the next release.