Triangulation Export
Allows you to automatically triangulate exports using a new tick option
that appears when, “Apply Modifiers” is active on a model, using option
defaults I think are the best for game exports.  If you still wish to
export triangulation manually however, ticking this option will not
overwrite any pre-existing triangulation modifiers you wish to apply
for export.

The next update will include the ability to export multiple objects as
a single FBX file, through Parent relationships and/or Groups.  This
requires a re-think of how parts of the UI and code work however, so I
don’t expect to release it too soon.