Group Export (Phase 2)
Yay new features!

- Unity Export Compatibility
Although i can’t test it because Unity won’t let me sign in to the damn
software atm, I implemented the code that should let it work.  I hope
this works?

- Separate Group Export Slots
Now as long as you’re using the auto-assign option, you can export
different object types as separate FBX files, as companions to the main
file export.  Use this to export high and low-poly versions of the same
object to bake maps in another program or separately export the
collision mesh for Unity.

- Better Group UI
I still need to work out some kinks in the UI, but for the time being i
improved it to be slicker and flow better.