Export Defaults (Phase 2)
Finalised the code to enable the exporting of assets in a job/pass
format, making it far more flexible and automated as well as easier to
apply the same export settings to multiple objects.

Animations still aren’t supported (yet), and are coming in the next
update.

Additionally, single objects can both now use the export pass system,
as well as the ability to use suffixes to define object components
(Low-Poly, Cage, etc).  The Auto-Assign option has now been removed, as
the plugin will automatically detect whether different components for
the same object exist or not (single object export only, coming soon
for groups).

Haven’t marked a release yet as it hasn’t been fully tested and bugs
are likely abound, so be warned when using this!