Auto-Assign Removal + Tweaks
Auto-Assign no longer has to be set by the user, as this feature seemed
needless and something that could be determined by code.

Additionally further tweaks were made to carry underlying code changes
to the group export code so it runs more efficiently and doesn’t export
any passes when it can’t collect any objects from the group that can be
exported for it.