Polish Ahoy! (Phase 1)
Added a bunch of nice little things to make the plugin all-round better

Single Object UI List
Although not feature complete, you can use this to inspect all the
objects in the current scene that are tagged for export.

Underlying Suffix Check Code
Instead of using a risky, basic str.find() function, im using something
more sophisticated to exactly identify suffix strings in object names,
to prevent an object that happened to use, “_LP” in the middle of it’s
name causing havoc on the export process

SUB-DIRECTORIES!
YES!  You can now name a sub-directory for any pass, which will be
created inside the specified Location Default if it doesn’t already
exist, and store all of the passes exports in that folder.