1.0 THINGS
- Added new plugin options to change whether the object list
auto-refreshes, as well as what secondary tool appears next to the
object lists.  You can choose between Focus and Select.
- Root objects can now be anything, or nothing at all. B)
- If an error is caught during the export process, your scene will
appear as if nothing ever happened to it.
- More than one group can now be edited in the scene at the same time,
using the new cursor button next to the list.  The same behaviour now
also applies to the Object tab, to standardise the experience.
- Fixed a bug where if the Sub-Directory had whitespace in it, the
process for getting the file path would wig out.
- Further UI refinements