1.0 - Final UI Changes (WIP 5)
- Changed a bunch of internal variables.  If someone is actually using
this plugin right now, you may see some objects lose variables.  Sorry!
- Fixed an issue when exporting objects in different scenes.  The
plugin will currently only export objects for the scene it currently is
seeing, although I hope to include support for full .blend file export
in the future.
- Fixed an issue when using a Blend Directory, if the file name had a
dot on it.
- Removed Asset List filtering options, due to their incompatibility
with some of the list features.  They will be re-introduced in a future
version.