1.0 - Full Unity/UE4 Export
- Despite Unity's best efforts, the Capsule plugin now supports seamless
exporting of Static and Animated objects.  There may be one or two bugs
left, but it should be working fine.

To get the full benefits of special export code in either, use one of
the special Saved Presets, as they enable special code to make exporting
in either much better.

- More tooltip renaming!