0.999.3 - Scene Preservation Bandaid #1
I’m working on a few last minute fixes to see if I can get a near 100%
success rate on changing the origins of complex objects or groups that
also have constraints and parent/child hierarchies.  Involves
constraint and parent hunting, and all manner of good times.

Theres a new tick in the Armature Preset menu called,  “Preserve
Armature Constraints”, which…. you can read the description for it, so
sleepy… z___z