Export - Improved Labels and Warnings
GLTF - NLA and Grouped Animation options are now better described and with useful defaults.
GLTF - A warning now appears when using Draco as it may not be clear to the user that it counts as a specialized GLTF extension.
USD - Fixed an incorrect description for USDZ as a file format.